TST: additional unit tests for Fragment class and inclusion of test_fragment_data_source file
